Week 6 Part 1: reorganize into src/features and src/shared; update imports
